==> notebooks/2. cleaning1.ipynb <==
***Issues***

- missing grape varieties in white_DF. Need to investigate how many and how to handle this.
- missing fizziness and tannin data expected as white wine have low tannin content and noone of the extracted wines are sparkling.
- remove fizziness from both datasets.
- 'ratings_count' and wine_ratings_count are very similar but with fewer counts for 'ratings_count'. same for 'averages' columns.

***Positives***
- appear to be very few null values in the datasets. Should further investigate to check.

This notebook aims to address any cleaning that needs to be completed before any analysis can take place.  
Cleaning will include:  
- Handling of missing/null values
- Correcting data types for each variable (column)
- Handling duplicate entries
- Remobal of unnecessary columns ('fizziness', 'ratings_count', 'ratings_average')
- Considering how to handle outliers

import json
import numpy as np
import pandas as pd

# load both datasets
white_wine_raw = pd.read_csv('../data/white_wine_raw.csv', sep='|')
red_wine_raw = pd.read_csv('../data/red_wine_raw.csv', sep='|')

# going to begin cleaning white wine df as believed to be more problematic
# quick inspection of df
white_wine_raw.info()

# not concerned about fizziness or tannin
# do need to look into grape_names once duplicates are dealt with.

# initially it looks good but need to check to duplicates...
white_duplicates = white_wine_raw[white_wine_raw.duplicated()]
len(white_duplicates)

# a lot of duplicates so created a new df without them. Lost A LOT of data
white_wine_temp = white_wine_raw.drop_duplicates(keep='first')
len(white_wine_temp)

white_wine_temp.info()

len(white_wine_temp[white_wine_temp['grape_names'].isna()])

whites = pd.read_csv('../white_clean2.csv',sep='|')

# a lot of missing grape names but im going to do it manually.
white_wine_temp.loc[(white_white_temp['vintage_name'].str.contains("Monestir Sant Sebastià dels Gorgs Blanco")) & (white_wine_temp['grape_names'].isna()), 'grape_names'] = 'Macabeo, Sauvignon Blanc'
#white_wine_temp.loc[white_wine_temp['wine_name'] == "Usatges Blanc", 'grape_names'] = 'Macabeo, Xarel-lo, Moscatel'
white_wine_temp.to_csv('white_wine_temp.csv', sep='|', index=False)

white_wine_temp[white_wine_temp['grape_names'].isna()]

# drop uneccessary columns
drop_cols = ['fizziness', 'ratings_count', 'ratings_average']
white_wine_temp.drop(columns=drop_cols, axis=1, inplace=True)

# add a marker for white wine for when i join the datasets
# and rename grape_names to grape_varieties
white_wine_temp['red/white'] = 'white'
white_wine_temp.rename(columns={'grape_names':'grape_varieties'}, inplace=True)

white_wine_temp.head()

# loaded a version i cleaned in another notebook
white_wine_temp = pd.read_csv('../white_clean2.csv', sep='|')

white_wine_temp.isna().sum()

white_wine_temp.to_csv('../data/white_wine_cleaned.csv', sep='|', index=False)

white_wine_temp.year.value_counts()

# start again with checking and cleaning red wine df
red_wine_raw.info()

# initially it looks good but need to check to duplicates...
red_duplicates = red_wine_raw[red_wine_raw.duplicated()]
len(red_duplicates)

# a lot of duplicates so created a new df without them. Lost A LOT of data
red_wine_temp = red_wine_raw.drop_duplicates(keep='first')
len(red_wine_temp)

red_wine_temp.info()

# got 2 nulls in year that need investigating
red_wine_temp.year.value_counts()

# Got 5 without values. Assuming 2 are NaNs and other 3 are something else!
# pull up NaNs
red_wine_temp[red_wine_temp['year'].isna()]

# look up and replace using...
red_wine_temp.loc[red_wine_temp['vintage_name'] == 'Bodegas Vilano Full Flap Red Blend', 'year'] = 2021

red_wine_temp['year'] = red_wine_temp['year'].astype(str)

red_wine_temp[red_wine_temp['year'] == 'N.V']

red_wine_temp.isna().sum()

# drop uneccessary columns
drop_cols = ['fizziness', 'ratings_count', 'ratings_average']
red_wine_temp.drop(columns=drop_cols, axis=1, inplace=True)

# add a marker for red wine for when i join the datasets
# and rename grape_names to grape_varieties
red_wine_temp['red/white'] = 'white'
red_wine_temp.rename(columns={'grape_names':'grape_varieties'}, inplace=True)

# final inspection before saving cleaned dataset
red_wine_temp.info()

red_wine_temp.to_csv('../data/red_wine_cleaned.csv', sep='|', index=False)
