==> notebooks/3. cleaning2.ipynb <==
import numpy as np
import pandas as pd
import functions as fn

white_wine = pd.read_csv('../data/white_wine_cleaned.csv', sep='|')
red_wine = pd.read_csv('../data/red_wine_cleaned.csv', sep='|')

white_wine.head()

red_wine.head()

# going to join the datasets
all_wine = pd.concat([white_wine, red_wine])

all_wine.head()

all_wine.info()

all_wine.to_csv('../data/all_wine_cleaned.csv', sep='|', index=False)
