==> notebooks/4. cardinality_transformations.ipynb <==
import numpy as np
import pandas as pd
import functions as fn

# load dataset
wine = pd.read_csv('../data/all_wine_cleaned.csv', sep='|')

# get cardinality of each column using function 'card_tipo'
# umbral_categoria = 10, umbral_continua = 30
cardinalityDF = fn.card_tipo(wine)
cardinalityDF

# got no categoric variables (except red/white but i made that so doesnt count)
# umbral_categoria = 24, umbral_continua = 30
cardinalityDF = fn.card_tipo(wine, 24, 30)
cardinalityDF

# i want a count for number of grapes in each wine so made a functions to use with apply
# still very few categoric variables (except red/white but i made that so doesnt count)
# going to do some transforming before moving into meaningful distribution analysis

wine['grape_count'] = wine['grape_varieties'].apply(fn.count_grapes)

# i also want to use regions in my analysis, so im going to investigate unique values:
wine.region.unique()

len(wine.region.unique())

# 91 unique values will be extremely messy so i will categorise into communidades
# there are three problem regions to investigate and correct: 
len(wine[(wine['region'] == 'Vino de España') | (wine['region'] == 'Vino de Pago') | (wine['region'] == 'Clàssic Penedès')])

wine[(wine['region'] == 'Vino de España') | (wine['region'] == 'Vino de Pago') | (wine['region'] == 'Clàssic Penedès')]

#code to change the regions to city closest to the bodega:
wine.loc[wine['bodega'].str.contains('Terras Mancas'),'region'] = 'Ourense'

'''
create a regions to communidades dictionary at the same time
also uses the regions from 'white_wine.region.unique()'
regions_dict = {
    'Andalucia':['Málaga', 'Cádiz', 'Sierras de Málaga','Montilla-Moriles','Ribera del Guadiana','Sierra Norte de Sevilla','Los Palacios','Andalucía','Sanlúcar de Barrameda','Jerez-Xérès-Sherry'],
    'Aragon':['Somontano','Valdejalón','Aragón','Cariñena','Calatayud','Ribera del Gállego-Cinco Villas','Campo de Borja','Aylés','Huesca'],
    'Asturias':['Asturias','Cangas'],
    'Baleares':['Mallorca','Binissalem-Mallorca','Plà i Llevant','Isla de Menorca','Formentera'],
    'Canarias':['Gran Canaria','Valle de la Orotava','Lanzarote','Valle de Güímar','Tacoronte-Acentejo','Ycoden-Daute-Isora','El Hierro','La Palma','Islas Canarias','Tenerife'],
    'Cantabria':['Liébana'],
    'Castilla y Leon':['Ribera del Duero','Castilla y León','Abadía Retuerta','Bierzo','Rueda','Toro','Tierra de León','Tierra del Vino de Zamora','Sierra de Salamanca','Arribes','Arlanza','Cigales','Valladolid','Zamora','Segovia','Sierra de Gredos','Salamanca','Burgos','Peñaranda'],
    'Castilla La Mancha':['Jumilla','La Mancha','Finca Élez','Dehesa del Carrizal','Méntrida','Valdepeñas','Uclés','Manchuela','Pago Calzadilla','Ribera del Júcar','Dominio de Valdepusa','Pago Florentino','Toledo','Ciudad Real','Guadalajara','Albacete','Cuenca'],
    'Catalunya':['Priorat','Empordà','Catalunya','Costers del Segre','Penedès','Terra Alta','Alella','Pla de Bages','Conca de Barberà','Montsant','Tarragona','Clàssic Penedès','Barcelona'],
    'Valencia':['Valencia','Alicante','Utiel-Requena','El Terrerazo','Villena'],
    'Extremadura':['Extremadura','Cáceres'],
    'Galicia':['Valdeorras','Rías Baixas','Ribeiro','Galicia','Ribeira Sacra','Monterrei','Barbanza e Iria','Ourense','Pontevedra','La Coruña','Val do Salnés'],
    'Madrid':['Madrid'],
    'Murcia':['Jumilla','Murcia','Almansa','Yecla','Bullas'],
    'Navarra':['Otazu','Navarra','Arínzano','Ribera del Queiles'],
    'Pais Vasco':['País Vasco','Bizkaiko Txakolina','Getariako Txakolina','Rioja Alavesa','Arabako Txakolina','Álava'],
    'La Rioja':['Rioja','Rioja Alta','Valles de Sadacia','La Rioja']
}
'''

# Cleaned regions in another notebook
wine = pd.read_csv('../wineDF_2.csv', sep='|')

wine.info()

# umbral_categoria = 24, umbral_continua = 30
cardinalityDF = fn.card_tipo(wine, 24)
cardinalityDF

wine.region.unique()

# now i can apply the function to get the community from each region:
wine['communidad'] = wine['region'].apply(fn.get_community)

# check for any missed regions that haven't found a communidad
wine[wine['communidad'] == 'Unknown']

# last quick inspection
wine.head(20)

# save final cleaned and ready dataset
wine.to_csv('../data/wine_df_final.csv', sep='|', index=False)

all_wine.grape_varieties.unique()

all_varieties = all_wine['grape_varieties'].str.split(',').explode().str.strip()

grapes_ordered = sorted(all_varieties)
